requirements.txt
flask
torch
torchvision
opencv-python-headless
Pillow
realesrgan
basicsr
numpy